The 1st Webpage
Here, I have a created the 1st webpage named Lannister Automobiles. As you can see clearly that I have created a webpage for a car dealership.
The image source can be of your choice, just download the image from internet and copy it into a folder and copy the path of the folder in img src.
The a href is used to link the other pages of your webiste.

(**Disclaimer :- I have written this code on my own it is neither copied from the internet nor created by the other person)